requirements.txt
pandas==2.1.0
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
cryptography==41.0.1 